+++ classes/games.rb
@@ -5,7 +5,7 @@ require './classes/author'
 require './classes/item'
 
 class Game < Item
-  attr_reader :multiplayer, :last_played_at
+  attr_reader :publish_date, :multiplayer, :last_played_at
 
   def initialize(multiplayer:, last_played_at:, publish_date: nil, archived: false)
     super(publish_date: publish_date, archived: archived)
