+++ data/authors.json
@@ -1,1 +1,1 @@
-{"id":723,"first_name":"le","last_name":"pa"}
+[{"id":994,"first_name":"Lod","last_name":"Osa"},{"id":11,"first_name":"Os","last_name":"Os"},{"id":61,"first_name":"JOhn","last_name":"Doe"}]
