+++ data/games.json
@@ -1,1 +1,1 @@
-[]
+[{"id":"058d361a-afe9-4ca6-83fa-79806ebb2b4e","publish_date":"2012-12-12","archived":false,"multiplayer":true,"last_played_at":"2011-11-11","author":{"id":61,"first_name":"JOhn","last_name":"Doe"}}]
