+++ data_manager.rb
@@ -12,7 +12,11 @@ class DataManager
   def self.load_games
     if File.exist?('./data/games.json')
       JSON.parse(File.read('./data/games.json')).map do |game|
-        Game.new(game['publish_date'], game['mode'], game['last_played_at'])
+        Game.new(
+          publish_date: game['publish_date'],
+          multiplayer: game['multiplayer'],
+          last_played_at: game['last_played_at']
+        )
       end
     else
       File.write('./data/games.json', JSON.dump([]))
@@ -23,7 +27,7 @@ class DataManager
   def self.load_authors
     if File.exist?('./data/authors.json')
       JSON.parse(File.read('./data/authors.json')).map do |author|
-        Author.new(author['id'], author['full_name'])
+        Author.new(first_name: author['first_name'], last_name: author['last_name'])
       end
     else
       File.write('./data/authors.json', JSON.dump([]))
